requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-extensions==3.0.3
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.1
idna==2.10
l18n==2018.5
Pillow==6.2.2
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9.2
wagtail-whoosh==0.2.4
webencodings==0.5.1
Whoosh==2.7.4
Willow==1.3
XlsxWriter==1.2.9
